Restructure for headless tests
